# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CognIA API
  version: 0.1.0
  description: |
    API backend de la tesis "Aplicativo web con Random Forest para la alerta temprana
    de cinco trastornos psicologicos en ninos de 6 a 11 anos".
    Proyecto academico del programa de Ingenieria de Sistemas y Computacion en la
    Universidad de Cundinamarca (grupo GISTFA).
    Esta API genera alertas de riesgo y no representa un diagnostico clinico.
    Trabaja solo con datos simulados o anonimizados.

    Enfoque de seguridad y buenas practicas:
    - Auth: JWT access token en JSON; refresh token en cookie HttpOnly con path
      /api/auth/refresh.
    - CSRF: double-submit con header X-CSRF-Token y cookie csrf_refresh_token; la
      cookie rota en refresh.
    - MFA: TOTP con secreto cifrado y recovery codes; obligatorio para roles
      privilegiados.
    - Observabilidad: /healthz (liveness), /readyz (DB readiness), /metrics (token
      opcional, por proceso).

    Notas:
    - Si usas cookies, define CORS_ORIGINS explicitos y envia credenciales.
    - /api/predict es experimental y puede cambiar.

    Creditos:
    - Andres Felipe Melo Chaguala (estudiante investigador)
    - Johan Thomas Cristancho Silva (estudiante investigador)
    - Oscar Jobany Gomez Ochoa (director)
    - Universidad de Cundinamarca, grupo GISTFA
servers:
  - url: /
    description: Same-origin (use the host that serves /docs)
  - url: http://127.0.0.1:5000
    description: Local development (IPv4)
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      description: |
        Creates a new user account. Username and email must be unique.
        Password must be at least 8 characters.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (password)
      description: |
        Authenticate with username and password.
        On success returns an access token in JSON and sets refresh_token and
        csrf_refresh_token cookies. If MFA is enabled, returns a challenge without
        cookies. If MFA enrollment is required, returns a short-lived enrollment
        token restricted to MFA setup/confirm. For valid credentials the endpoint
        always responds 200, with one of:
        - access_token + cookies (login complete)
        - mfa_required + challenge_id (no cookies)
        - mfa_enrollment_required + enrollment_token (no cookies)
        Rate limit: 5 per 15 minutes per IP.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access token or MFA challenge
          headers:
            Set-Cookie:
              schema:
                type: string
              description: refresh_token (HttpOnly) + csrf_refresh_token
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthTokenResponse"
                  - $ref: "#/components/schemas/MFAChallengeResponse"
                  - $ref: "#/components/schemas/MFAEnrollmentResponse"
              examples:
                token:
                  summary: Access token response
                  value:
                    access_token: "<jwt>"
                    token_type: bearer
                    expires_in: 900
                mfa_challenge:
                  summary: MFA challenge response
                  value:
                    mfa_required: true
                    challenge_id: "<uuid>"
                    expires_in: 300
                    error: mfa_required
                mfa_enrollment:
                  summary: MFA enrollment response
                  value:
                    mfa_enrollment_required: true
                    enrollment_token: "<jwt>"
                    token_type: bearer
                    expires_in: 600
                    msg: MFA enrollment required
                    error: mfa_enrollment_required
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/auth/login/mfa:
    post:
      tags: [Auth]
      summary: Complete MFA login
      description: |
        Completes the MFA challenge using a TOTP code or a recovery code.
        If MFA enrollment is still required, returns 403 with
        error mfa_enrollment_required.
        Rate limit: 3 per 10 minutes per IP.
      operationId: loginMfa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginMFARequest"
      responses:
        "200":
          description: Access token
          headers:
            Set-Cookie:
              schema:
                type: string
              description: refresh_token (HttpOnly) + csrf_refresh_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Uses the refresh_token cookie to issue a new access token.
        Requires a CSRF double-submit: X-CSRF-Token header must match the
        csrf_refresh_token cookie. Refresh token is rotated on every call.
        If MFA enrollment is required, returns 401 and clears auth cookies.
      operationId: refresh
      security:
        - cookieAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: New access token
          headers:
            Set-Cookie:
              schema:
                type: string
              description: refresh_token rotated + csrf_refresh_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CsrfFailed"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh tokens)
      description: |
        Requires an access token plus CSRF header and cookie.
        Revokes all refresh tokens for the user.
      operationId: logout
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: Logged out
          headers:
            Set-Cookie:
              schema:
                type: string
              description: auth cookies cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
components:
  parameters:
    CsrfHeader:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
      description: Must match the csrf_refresh_token cookie value.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token sent in the Authorization header.
    mfaEnrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived token for MFA setup/confirm only (claim mfa_enrollment=true).
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token cookie (HttpOnly).
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_refresh_token
      description: CSRF cookie used for double-submit validation.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF header used for double-submit validation.
    metricsToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer token for /metrics (Authorization: Bearer <token>)."
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
        error:
          type: string
      example:
        msg: Invalid credentials
        error: invalid_credentials
...
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: testuser
        email:
          type: string
          example: test@example.com
        password:
          type: string
          format: password
          example: P4ssw0rd!
        full_name:
          type: string
          example: Test User
    RegisterResponse:
      type: object
      properties:
        msg:
          type: string
          example: user created
        user_id:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: testuser
        password:
          type: string
          format: password
          example: P4ssw0rd!
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 900
    MFAChallengeResponse:
      type: object
      properties:
        mfa_required:
          type: boolean
          example: true
        challenge_id:
          type: string
          format: uuid
        expires_in:
          type: integer
          example: 300
        error:
          type: string
          example: mfa_required
    MFAEnrollmentResponse:
      type: object
      properties:
        mfa_enrollment_required:
          type: boolean
          example: true
        enrollment_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 600
        msg:
          type: string
          example: MFA enrollment required
        error:
          type: string
          example: mfa_enrollment_required
    LoginMFARequest:
      type: object
      description: Provide either code or recovery_code.
      required: [challenge_id]
      properties:
        challenge_id:
          type: string
          format: uuid
        code:
          type: string
          description: TOTP code
          example: "123456"
        recovery_code:
          type: string
          description: One-time recovery code
          example: "a1b2c3d4"
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
          example: logged out
...